Update DPA with new UKS key
UKS key schema facilitate operations:

- PUT: a key is generate for each object. That key is sent back to S3
once the object is successfully uploaded (and protected by erasure
codes).
Each fragment of the object is identified by an other key that is
generated based on the object's key and its index in erasure codes.
Fragment keys are not stored as they can be regenerated when the
object's key is known.

- GET: given an object's key, keys of its fragments are generated
in order to retrieve them from data storage backend.

- DELETE: given an object's key, keys of its fragments are generated
in order to delete them from data storage backend.